+++ ginga/gw/Readout.py
@@ -38,7 +38,7 @@ class Readout(object):
         self.maxy = 0
         self.maxv = 0
 
-        self.fitsimage = None
+        self.ext_viewer = None
 
     def get_widget(self):
         return self.readout